requirements.txt
torch==1.13.0
torchvision==0.14.0
tensorflow==2.11.0
pandas
seaborn
pillow
matplotlib
scikit-learn
argparse
opencv-python
argparse
dicom2jpg
nibabel